Fix clang compile error
